feat: add heading blocks
