+++ src/content/blog/test.mdx
@@ -8,6 +8,14 @@ title: Execution Context in JavaScript
 publishedDate: 2023-11-29
 ---
 
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Odio sint aliquam iste fugiat iure pariatur fugit
+inventore alias, ut totam distinctio, illum magni magnam modi.
+
+## What is Execution Context
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Odio sint aliquam iste fugiat iure pariatur fugit
+inventore alias, ut totam distinctio, illum magni magnam modi.
+
 This is sample content for the blogs.
 
 This is another sample.
@@ -15,6 +23,21 @@ This is another sample.
 Adding **bold** and _italic styles_ and _**bold italic styles**_
 
 Then we also have links like these [Go to home](/ 'Hello World')
+
+## Phases of Execution Context
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Odio sint aliquam iste fugiat iure pariatur fugit
+inventore alias, ut totam distinctio, illum magni magnam modi.
+
+### Memory Creation Phases
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Odio sint aliquam iste fugiat iure pariatur fugit
+inventore alias, ut totam distinctio, illum magni magnam modi.
+
+### Code Execution Phase
+
+Lorem ipsum dolor sit amet consectetur adipisicing elit. Odio sint aliquam iste fugiat iure pariatur fugit
+inventore alias, ut totam distinctio, illum magni magnam modi.
 
 <BlockQuote
 	citationSource="https://en.wikipedia.org/wiki/JavaScript"
